src/requirements.txt
boto3==1.26.19
rasterio==1.3.3
numpy==1.23.4
Pillow==9.3.0
